+++ R/drawCellShiny.R
@@ -1,6 +1,11 @@
 #' Run Shiny app for the drawCell Package
 #'
 #' @export
+#'
+#' @import shiny
+#' @import shiny.semantic
+#' @import htmltools
+#' @import glue
 drawCellShiny <- function(){
     shiny::shinyAppDir(appDir = base::system.file('shinyApp/drawCellShiny',
                                                   package = 'drawCell'))