+++ NEWS.md
@@ -1,4 +1,5 @@
-# drawCell 0.99.0
+# drawCell (development version)
 
 
 * Submission to Bioconductor 3.14
+* Improve shiny app with `shiny.semantic`